+++ tests/date-fns/services/dateFnsDateApi.service.test.ts
@@ -1,5 +1,5 @@
-import {DateApi, DateTimeRelativeParser} from '@anglr/datetime';
-import {DateFnsDateApi} from '@anglr/datetime/date-fns';
+import {DateApi, DateTimeRelativeParser, FormatProvider} from '@anglr/datetime';
+import {DateFnsDateApi, DATEFNS_FORMAT_PROVIDER} from '@anglr/datetime/date-fns';
 import {sk} from 'date-fns/locale';
 
 import {dateApiTests} from '../../src/services/dateApi.tests';
@@ -28,18 +28,35 @@ function initialize(): DateApi<Date>
 describe('DateFnsDateApi class', () =>
 {
     let dateApi: DateApi<Date> = null!;
+    let formatProvider: FormatProvider;
 
     beforeAll(() =>
     {
         dateApi = initialize();
+        formatProvider = DATEFNS_FORMAT_PROVIDER.useFactory();
     });
 
-    it('weekStartsOnMonday method', () =>
+    test('weekStartsOnMonday method', () =>
     {
         expect(dateApi.weekStartsOnMonday()).toBe(true);
     });
 
-    it('weekdaysShort method', () =>
+    test('getFormat method => date', () =>
+    {
+        expect(dateApi.getFormat(formatProvider.date)).toBe('d. M. y');
+    });
+
+    test('getFormat method => time', () =>
+    {
+        expect(dateApi.getFormat(formatProvider.time)).toBe('H:mm');
+    });
+
+    test('getFormat method => dateTime', () =>
+    {
+        expect(dateApi.getFormat(formatProvider.dateTime)).toBe('d. M. y H:mm');
+    });
+
+    test('weekdaysShort method', () =>
     {
         expect(dateApi.weekdaysShort()).toEqual(['po', 'ut', 'st', 'št', 'pi', 'so', 'ne']);
     });