+++ tests/moment/services/momentDateApi.service.test.ts
@@ -1,5 +1,5 @@
-import {DateApi, DateTimeRelativeParser} from '@anglr/datetime';
-import {MomentDateApi} from '@anglr/datetime/moment';
+import {DateApi, DateTimeRelativeParser, FormatProvider} from '@anglr/datetime';
+import {MomentDateApi, MOMENT_FORMAT_PROVIDER} from '@anglr/datetime/moment';
 import moment from 'moment';
 
 import {dateApiTests} from '../../src/services/dateApi.tests';
@@ -27,18 +27,35 @@ function initialize(): DateApi<moment.Moment>
 describe('MomentDateApi class', () =>
 {
     let dateApi: DateApi<moment.Moment> = null!;
+    let formatProvider: FormatProvider;
 
     beforeAll(() =>
     {
         dateApi = initialize();
+        formatProvider = MOMENT_FORMAT_PROVIDER.useFactory();
     });
 
-    it('weekStartsOnMonday method', () =>
+    test('weekStartsOnMonday method', () =>
     {
         expect(dateApi.weekStartsOnMonday()).toBe(true);
     });
 
-    it('weekdaysShort method', () =>
+    test('getFormat method => date', () =>
+    {
+        expect(dateApi.getFormat(formatProvider.date)).toBe('DD.MM.YYYY');
+    });
+
+    test('getFormat method => time', () =>
+    {
+        expect(dateApi.getFormat(formatProvider.time)).toBe('H:mm');
+    });
+
+    test('getFormat method => dateTime', () =>
+    {
+        expect(dateApi.getFormat(formatProvider.dateTime)).toBe('DD.MM.YYYY H:mm');
+    });
+
+    test('weekdaysShort method', () =>
     {
         expect(dateApi.weekdaysShort()).toEqual(['po', 'ut', 'st', 'št', 'pi', 'so', 'ne']);
     });